Clothing recommendation related classes moved to appropriate package.
